+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Fixed character position starting at normalized CRLF (#786)
 - Improved setting `Href` in `Url`
 - Fixed crash for invalid attribute names during SVG parsing (#795)
+- Added more punycode replacement characters (#797)
 
 # 0.12.1